Replaced enum for ElementType with structs
The advantage is to collect all in the same place the info that a specific element type needs to provide. Also, since we are using templates, there is no need for, say, TRI and TET to be of the same type but just to be able to provide specific information (e.g. physicalDim, parametricDim, ...).