math and quadrature: Redesigned quadrature rule
QuadratureRule is a template class with respect to QuadratureType (e.g. GAUSS, ...), ElementType (e.g. TRI, TET, ...), and the degree of exactness r.
The Integrator owns a static const instance of the quadrature rule (i.e. a tuple of points and weights) which is known at compile time.
Work in progress, this breaks the test math/tests/main.cc because the methods Integrator::_computeQuadPointCoordinates and Integrator::integrate have temporarily been commented out.

Co-authored-by: Anwar Koshakji <anwar.koshakji@gmail.com>