Removed notion of physical dim for ElementType
ElementType should not be aware of the dimension of physical space, as it has the meaning of 'reference element'.
The notion of the dimension of physical space should be owned by the specific element that is built on top of that ElementType.
For example, we might want to use the same reference element (and thus the same quadrature rule) for triangles in 2D or 3D.