mito/mesh: Simplified registration of entities compositions in mesh object
The information on the composition of the entity given in input to _insertEntityComposition was redundant, as the Simplex object (i.e. the entity to be inserted) has that information and can provide it within the method _insertEntityComposition.
However, this required to switch in mito/mesh/simplex.h the attribute _entities from an array of reference wrappers to an array of pointers for consistency with the data structures in the mesh object.

Also, the compositions sorting, which was done in method _insertEntityComposition to guarantee uniqueness of inserted entitites with respect to their composition, is now done in the contructor of each Simplex, so that two simplices having the same composition will result in two identical instances of class Simplex.