mito: Major improvements of Field class
Before this commit, class Field represented a scalar field. Fields were then aggregated to compose VectorFields. A better design is to use the template return type Y of Function<X, Y> to represent reals, vector, tensors, ...  and make Field general with respect to whether the field is scalar, vector, or tensor. In the current implementation, Field provides the interface for both ScalarField and VectorField (and the future TensorField), as
ScalarField<D> is simply Field<D, real> and VectorField<D, N> is Field<D, vector<N>>.
The integrator now takes in input a Field and can integrate scalar, vector, and tensor fields seamlessly.