mito/mesh: {SimplexFactory} now stores {simplex_ptr} instead of {simplex_t}
The implementation of {SimplexFactory} is now consistent with that {OrientedSimplexFactory}, which also stores shared pointers to oriented simplices.
This also allows major simplifications in the implementation of static method {orientedSimplex}. In fact, now the {OrientedSimplexFactory} does not need to check for the existence of an oriented simplex with opposite orientation to fetch the simplex footprint.