segmented-container: add prototype of segmented container
The segmented container is a resizable container stable in memory.
This container is implemented as a vector of pointers to segments of memory of size N (N being a template parameter of the data structure). When the container runs out of space, a new segment of memory is allocated and added to the pile.

Co-authored-by: Anwar Koshakji <anwar.koshakji@gmail.com>
Co-authored-by: Michael Aivazis <michael.aivazis@para-sim.com>