lib/mito/topology: remove redundancy of {simplex} and {footprint} accessors of {OrientedSimplex}
The {footprint} is the accessor for the underlying (unoriented) {Simplex} and returns a reference to an {unoriented_simplex_ptr<D>} instance.

The {simplex} accessor was only necessary to access the true immutable address of the {OrientedSimplex} instance (as opposed to the address of the shared pointer that wraps it) to serve as a unique identifier.
The fact that we use the simplices address as unique identifiers is an implementation detail that is now hidden below the method {simplex_id}. Similarly, this commit introduces a {footprint_id}, returning the unique identifier (i.e. the address) of the (oriented) simplex's (unoriented) footprint.