lib/mito/utilities: fix {SegmentedContainerIterator::operator++}
*Problem
When an element was added at the end of the container, and therefore
a new segment was created, the iterator did not know how to point to
the new segment.

*Solution
The iterator now points to the new segment when the end of the segment
is reached but not the end of the container.

Co-authored-by: Bianca Giovanardi <bianca.giovanardi@gmail.com>