lib/mito/utilities: remove redundant {_data} attribute from {SegmentedContainer}
This attribute was meant to store the pointers to the beginning of each segment.
However, after rev. fa25f94, each segment stores a hint to the location of the next segment, so the {_data} attribute has become redundant.
The only necessary information carried by {_data} is:
- the beginning of the container
- the end of the last segment
- the number of segments
These attributes have been added in replacement of the more expensive and complex {_data} data structure.