lib/mito/topology: temporarily make constructors of simplices public
These constructors cannot be public because only the appropriate factory has 'construction privileges' (in fact the factory is responsible for making sure that each simplex is unique).
However, the simplex factory needs the segmented container to instantiate the simplex, so the segmented container would need to call the simplex constructor. In the meantime that we figure out whether the segmented container should be befriended with the oriented simplex or rather should the oriented simplex factory place instantiate the simplex at an available location in the segmented container, I make the constructor public.