tests/mito.lib/shared-pointer: fix {TestSegmentAllocation}
This test failed to build with {gcc12} with error 'pointer used after free'.
In fact, the destructor of the {shared_ptr} to {resource} was being called after the deallocation of all the segment memory.
Therefore, the reference count of {resource} was attempted to be accessed and modified after the memory of {resource} was freed.
This has been fixed by doing the allocation/deallocation of memory within a static class, whose destructor is called after all regular destructors. This is also a more representative case to test, because the end use of the {Shareable} and {SharedPointer} classes in mito is in combination with the (static) simplex factories.