lib/mito/utilities: add method {clear} that deallocates the memory in {SegmentedContainer}
This resolves the memory errors in the destructors of the simplex factories arising from the random order of deletion of the simplices collection (static) attribute and the relative (static) factory map. In fact, the factory maps (i.e. {_orientations} and {_compositions} for {OrientedSimplexFactory} and {SimplexFactory}, respectively) need to be deleted before the simplices collection (i.e. {_oriented_simplices} and {_simplices}). Should this not be the case, the destructor of the map, which calls the destructor of a {mito::shared_ptr}, will try to decrease the reference count of the resource, which is owned by the resource and allocated by the segmented container, resulting in a memory error.

However, the current implementation breaks the simple case that the segmented container is instantiated in the main function (see tests/mito.lib/segmented-container).

One way to fix it is to give up the {static}-ness of attributes in the factories. Work in progress.

Co-authored-by: Anwar Koshakji <anwar.koshakji@gmail.com>