lib/mito/topology: major redesign of {topology}
Because of memory errors, it was not practical to maintain {OrientedSimplexFactory} and {SimplexFactory} as static classes (see rev. b891104 for a description of the memory errors). In the new design, {Topology} is a collection of {OrientedSimplexFactory} instances of different dimensions. Each of these factories, in turn, owns and manages a {SimplexFactory}.

See tests/mito.lib/mesh/main.cc for how topological entities are now to be instantiated.

This revision breaks several tests. Work in progress includes:
- redesign the point cloud based on a segmented container of points and make it nonstatic
- adjust all tests to the new design (each test will instantiate a topology and/or a point cloud as needed)
- fix the hierarchical erase of elements (and points!)
- adjust the mesh readers