lib/mito/utilities: temporarily remove {requires} clause in {SegmentedContainer}
This clause requires that the resource stored is 'reference countable'. However, because the compiler needs to see the header file to evaluate the constraint, there is a circular reference of header files that is not easy to solve.

This will be addressed in a separate issue, see also rev. ca79cd9.