lib/mito/topology: vertices are now treated equally to other oriented simplices
Class {OrientedSimplexFactory<0>} keeps a copy of the shared pointer to vertices that it instantiates, so as to be able to do reference tracking the same way the other {OrientedSimplexFactory}'s do. Now vertices are also returned by reference to {vertex_t} as opposed to by value.

This allows test {TestEraseElementTopology} to the further check that vertices that are not shared by an oriented simplex any longer are erased from {OrientedSimplexFactory<0>}.