lib/mito/topology: major refactoring of headers
The {oriented_simplex_ptr} {typedef} needs to see the header {OrientedSimplex.h} so as to verify the {ReferenceCountedObject} concept. On the other hand, class {Simplex} needs to see the {oriented_simplex_ptr} in a forward declaration.

For the reason above, the forward declarations have been split into two files.

Also, the free functions in {OrientedSimplex.h} have moved to a {utilities.h} header, as they need to see forward declarations of both {forward1.h} and {forward2.h}.