tests/mito.lib: tests now assemble triangles in a head-tail connected fashion
This is to force the client to construct the oriented simplices with correct orientation.
In fact, it is important that internal boundary simplices occur in the topology two times (one per orientation), as the criterion to establish whether an oriented simplex is on the boundary or not is whether the oriented simplex occurs once or twice in the topology.