lib/mito/mesh: major redesign of class {Mesh}
Class {Mesh} is now only template with respect to the cell type (e.g. triangle_t) and the dimension of the physical space (e.g 3 for 3D).
An instance of class {Mesh} does not store  the subcells any longer.
For example, an instance of class {Mesh<triangle_t, 3>} now only stores the triangular cells and not the segments or the vertices.
This greatly simplifies the mesh object. The lower dimensional cells can still easily be retrieved from the higher dimensional cells.