lib/mito/topology: vertices are now implemented as regular oriented simplices
The trivial case for the recursion in the simplices definition is now shifted to {Simplex} and {SimplexFactory}.
The motivation for having a vertex orientation is as follows: say that a segment goes from a (-) vertex to a (+) vertex (or viceversa), then the boundary identification algorithm extends seamlessly to 1D (a vertex is on the boundary if it appears in the mesh with only one orientation).
Giving vertices an orientation may also be useful in the 1D DG implementation.