lib/mito/utilities: available locations in {SegmentedContainer} are now filled in place
In {SegmentedContainer} implemented mechanism for which the available locations for element (over)writing are written directly into the spare locations of the segment, so that each segment hole stores a pointer to the next segment hole.

This mechanism is not compatible with an iteratable container, as writing pointers in place on the invalid elements causes method {isValid} of the stored resource to give wrong results. Therefore, iterators are not able to skip invalid elements any longer.

Co-authored-by: Bianca Giovanardi <bianca.giovanardi@gmail.com>