lib/mito: use a static variable for the resource of {Singleton} class
Avoid the use of a pointer since it prevented the destructors to be called.
Remove inheritance of singleton classes from {Singleton}.
Remove test fixture for cleanup the singletons. Each test that uses
a singleton needs to have a separate executable in order not to reuse
singletons among tests. This will be done in a follow-up commit.