Add {cmake} function for test cases that produce some output that needs to be checked
For the test cases that produce output (e.g. a {vtk} file), an
additional test is appended which checks the generated output using
{pytest}. The python script must be placed in the same directory of the
test and the file must have the same name of the test file.
Since for {io} tests have been split in separate files, each one has
now its own python script to check the output of the {vtk} writer.
Files are now cleaned up at the end of the python script.