lib/mito/topology: fix major bug in factories destructors
The order of the destructors of the factories of simplices in {Topology}
was undefined because so is the order of destructions of entries in
a {std::tuple}. This was causing memory leaks due to sometimes the
vertices factory being destroyed before the higher order simplex
factories.

The tuple has been split in different attributes and the order of
destruction is now deterministic.

Co-authored-by: Bianca Giovanardi <bianca.giovanardi@gmail.com>