lib/mito: explicitly call destructors of {Shareable}s
{Shareable} items are meant to be constructed by {SegmentedContainer}
by means of a placement new. A placement new decouples the memory
allocation from the instantiation of the object. Therefore, by doing
placement new, the destructor of the object must be called explicitly.
Because the factory instantiate {Shareable} (at the memory position
supplied by {SegmentedContainer}), they should also explicitly call
the destructors of the {Shareable} objects.