lib/mito/topology: add method to check if a segment exists based on its vertices
Add methods to find unoriented and oriented simplices based on a
given composition.

In {tests.mito.lib.flip-diagonal}, modify the asserts so that {segment_e}
and {segment_e_flip} are not used at the end of the test to check whether
they have been erased. Since in the test the segments are taken by
reference and they get deleted in {mesh::flipDiagonal},
checking them at the end of the test produced undefined behavior. Now we
can ask the topology to check whether they exists based on their vertices.