tests/mito/erase-element: remove obsolete part of test
The test removed (which was previously commented out) was meant to test the erasion of a simplex in an isolated environment from the simplex factories.
However, this test does not work any longer are simplices cannot be instantiated directly (they have a private constructor).
Furthermore, the erasion algorithm is the one that ended up being implemented in the topology.

Co-authored-by: Bianca Giovanardi <bianca.giovanardi@gmail.com>