tests/mito.ib/erase-element: minor improvement in test
Added more checks and improved schematics.

Copying simplices by value is cheap as the copy is shallow. Also, this is much safer as now the driver use of the simplex is included in the reference count. This means that it is now impossible to erase simplices in used by the driver.

Co-authored-by: Bianca Giovanardi <bianca.giovanardi@gmail.com>