benchmarks: add {mito} integration test
This test calls all the main {mito} functions/classes.
The test instantiates a {Mesh} using a {Topology}, a {Geometry} and a {PointCloud}.
Then it flips the common edge of the two triangles and uses {Tetra} on
the resulting mesh. Finally a {Manifold} is created and a cosine function
is integrated on it.

This will serve as benchmark test to run valgrind.