lib/mito/mesh: method {erase} now does not erase the element from the {Geometry}
This way, the {Geometry} attribute of class {Mesh} can be made {const}.
This is consistent with the fact that mesh cells are not registered by the mesh with the geometry, but rather first created and then inserted to the mesh. Similarly, the mesh should not attempt to erase cells from the geometry upon cell erasion in the mesh.