lib/mito/mesh/metis: partitioner now works correctly with multiprocessing
This commit fixes test {tests.mito.lib.mesh.metis_partitioner_mpi_load_mesh}.

Because the order of cells in the mesh (and that of vertices in a cell) is decided at runtime depending on where they are allocated in memory, the element connectivity datastructure fed to the {metis} partitioner may be different (although equivalent) from a run to one another.
When partitioning a mesh with multiprocessing, the different processors running concurrently may end up with different mesh partitions, leading to problems in communication.

This issue is solved by this commit by reordering geometrically (lexicographically) the cells within the mesh and the vertices within the cells in the mesh partitioner function.

Another way to solve this issue could have been to read the mesh on processor 0 and then broadcast it to everyone else, as suggested by Michael. Because this option requires to broadcast topological and geometrical information and datastructures, much like in a {parmetis} implementation, this option will be explored when we implement {parmetis} support.