lib/mito/geometry: coordinates of {Point} are now {const}
At this point of the design, I am not sure if points should be constants or not.
For example, would it make sense to envision that a point would move in space (as opposed to an object occupying one point would move to occupy another point)?
In the meantime, I pick the conservative version: points do not move.