lib/mito/topology: the simplex factories now call the destructor of simplices upon simplex erasure
Note that this destructor call does not free the memory, because the simplices were constructed with placement new.
The call to the destructor upon erasure is symmetric with the fact that the simplex factory also call the constructor of the simplices.
Before this commit destructors of deleted simplices were never called.

This new implementation makes  the simplex {_erase} method redundant as that method was de facto resetting the state of the simplex, which is not needed any longer as the simplex is now being destroyed.
Also, with this new implementation it is now possible to mark the attributes of simplices as {const}, which was not possible before due to the need of resetting simplices.