lib/mito/topology and lib/mito/utilities: improve workflow in element erasure
The callback mechanism implemented in rev. 1538ba04 (in which a shared pointer would notify the segmented container when the last shared pointer wrapping a given resource would go out of scope) is not actually needed, because the factories always hold the last instance of elements.
Therefore in this new implementation, the factories call the erase method of {SegmentedContainer}, as opposed to the shared pointer. This new implementation also greatly simplifies class {SharedPointer}, which no longer needs a pointer to the segmented container where the resource is stored.