lib/mito: major redesign of the {mito} utilities
The design of the {SharedPointer} and {SegmentedContainer} has been completely revisited.
The main motivation for this redesign is the ugliness resulting from the {SharedPointer} keeping a (non const!) pointer to to the {SegmentedContainer}, which caused {const} segmented containers not to be iterable.

Following rev. 60d9685, the new design does not delegate to the {SharedPointer} the responsibility of notifying the {SegmentedContainer} of an object being unused, like it was done before. Instead, a new class, the {Repository}, is now responsible for the creation and the destruction of resources.
The {Repository} class maintains the resources collection in a {SegmentedContainer}, therefore allowing a more net separation of the handling of the memory (which is the responsibility of the {SegmentedContainer}) from the instantiation of the resources ({Repository}).