lib/mito: redesign of simplex factories and point cloud
After this revision, the simplex factories and the point cloud do not own an instance of the shared pointers to their resources any longer.
The purpose of the former implementation was for the factory to always hand to the client the same instance of the shared pointer so as to avoid losing track of the resource reference count. However, because for a {Shareable} resource the reference count is stored together with the resource itself, a (regular) pointer to resource can simply be wrapped in a shared pointer at any time without losing track of the reference count.

The new implementation significantly simplifies reference counting on the client side, as no 'hidden instances' contribute to the reference count.