lib/mito/utilities: simplified {SegmentedContainer} iterators
In the meantime that we figure out what {const} iterators should look like, we remove the {const} iterators all together.

The current implementation does not work as intended. In fact, dereferencing a {const} iterator returns a shared pointer to a {const} resource, which is in general of different type with respect to the type passed as template argument to the {SegmentedContainer}.
For the time being, we use the {SegmentedContainer} data structure only for immutable types (all attributes of simplices and points are {const}).

If, in the future, we will need to use this data structure for a non {const} resource, we will revisit the design of these utilities.

Co-authored-by: Anwar Koshakji <anwar.koshakji@gmail.com>