lib/mito/topology and mesh: improve corner case of simplices being erased twice
Simplices can now be erased twice from a {Topology} or a {Mesh}.

When a simplex is erased from {Topology}, its shared pointer is reset to {nullptr}. Method {Topology::erase} checks that the simplex being erased is valid (i.e. not {nullptr}) and, if it is not, then the method does nothing.

Also, {Mesh::erase} does nothing if the cell in argument is not among the mesh cells.