.vscode: add configuration to be able to debug tests by using {vscode} integration with {cmake}
To debug a test, the {cmake} target for the test can be selected from
{vscode} toolbar.
For {linux}, the test can then be run directly from the toolbar.
For {macos}, there is a bug in {vscode} for which {lldb} cannot be found
if launched by the toolbar. Therefore, an additional configuration has
been added to {launch.json} in order run and debug a test that has been
selected as launch target in {vscode} toolbar.