lib/mito/manifolds: major bugfix in cell volume calculation
The volume of a simplex is calculated as 1/N * w(point)(v1, v2),
where {w} is the volume form, which is contracted with {v1, v2} the two directors of the simplex.
The point that the volume form needs to be evaluated at is not the quadrature point, but the the vertex that the two directors {v1} and {v2} stem from.