lib/mito/utilities: bugfix in {SegmentedAllocator} move constructor
This bug caused the move constructor of {SegmentedVector} to be faulty.
The default move constructor of {SegmentedAllocator} did not invalidate the source, which was still iterable. Therefore, the {SegmentedVector} destructor was able to call destructors of resources for both the source and the target {SegmentedAllocator} instance, resulting in destructors being called twice for the same instance of resource.

Add test to check the correct behavior of the {SegmentedVector} move constructor.