lib/mito/fem: bugfix
Fixed bug for which nodes in {NodalField} were duplicated across adjacent elements. When nodes are inserted in a mesh with mesh.insert({node0, node1, node2}), nodes are copied. Therefore, a hash function and {operator==} based on the nodes addresses will create copies for each instance of a node in a mesh (that is for the same node appearing in adjacent elements).

For now, the problem is fixed by setting nodes equal if their underlying vertex and points are equal and by editing the hash function to trivially combine the hashes of vertex and point (this needs to be improved to reduce the chance of clash).

However, these choices will be revisited when we implement discontinuous Galerkin.