lib/mito/io/vtk: {VTKFieldWriter} class redesign
The class now takes {gridWriterT} (i.e. mesh writer vs. cloud writer) instead of {gridT} (i.e. mesh vs. grid) as a template parameter. This will simplify the generalization of the class to parallel {vtk} writers.

Also, this commits clarifies at the api level the distinction between a {grid_writer} (which only writes mesh or point cloud) and a {field_writer} (which provides the capability to attach fields to meshes and point clouds).