lib/mito/utilities: fix memory leak in {SegmentedVector}
The memory leak occurred upon the reuse of memory locations of erased elements. In fact, method {erase} did not call the resource destructor upon element erasure, while the {emplace} would construct a new resource at the same location of the prevoiusly erased (but not destructed) resource.

One possible fix would be to call the resource's destructor in the {erase} method. However, because iterators on this data structure need to access information on the resource validity (so as to skip invalid elements), even resources marked as invalid (i.e. the erased resources) need to be able to report on their state of validity, which cannot happen after the resource destructor is called.

The solution implemented in this commit avoids the issue by calling the old resource destructor in method {emplace}, right before calling the constructor of the newly emplaced resource.

This implementation required the {SegmentVector} to be informed by method {location_for_placement} on whether the location for placement of the new resource is a reused location or not. This information is passed in the return value of method {location_for_placement}.