solvers: add methods {initialize} and {finalize} to class {PETScKrylovSolver}
Method {finalize} is needed as, in case {petsc} was initialized outside of the {PETScKrylovSolver} class, calling in the class destructor the {petsc} functions that free the memory results in a runtime error. In fact, in that case, the function {PetscFinalize} would get called before the destructor calls (and no {petsc} function can be called after {PetscFinalize}).

Method {inizialize} is implemented for the sake of symmetry with method {finalize}.