geometry: bugfix in {geometric_simplex} factories
Before this commit, the methods that build geometric simplices on top of a given simplex and a collection of nodes made copies of the nodes to be stored in the geometric simplex. The unwanted behaviour was causing a bug in the extraction of the boundary simplices of a mesh as it led to the duplication of all the boundary nodes in the boundary simplices. This has now been fixed.
The methods that build geometric simplices on top of a given simplex and a collection of nodes do not make copies of nodes any more.