discretization: major redesign of {FunctionSpace} based on the concept of discretization node
This commit introduces the notion of {DiscretizationNode} as separate from mesh node. While a mesh node is a pairing of a mesh vertex with a point, a {DiscretizationNode} may or may not correspond to a vertex. For example, in higher order discretizations, we have discretization nodes internal to (sub)simplices. Also, in case of discontinuous finite elements, we may have several discretization nodes for each mesh node.

Class {DiscretizationNode} is for now an empty class, only used to store the node label (which corresponds to the memory address of the node instance).