solvers/cuda/CUDADenseSolver: hardcoded mito::real to double to avoid compilation error
For some reason, when the CUDADenseSolver.cu is compiled with the nvcc compiler, it fails to compile with the error:

```
/usr/local/pyre/include/pyre/tensor/traits.h:30:18: error: expected identifier before ‘>’ token
   30 |         template <int N>
      |                  ^
/usr/local/pyre/include/pyre/tensor/traits.h:41:18: error: expected identifier before ‘>’ token
   41 |         template <int N>
      |                  ^
/usr/local/pyre/include/pyre/tensor/algebra.icc:1232:50: error: expected identifier before ‘>’ token
 1232 | template <pyre::tensor::square_matrix_c matrixT>
      |                                                  ^
```

Apparently, the nvcc compiler is not able to compile the some template code in the pyre library which we are using in the CUDADenseSolver.cu to defined the real type. This could be because nvcc doesn't support all the features of the C++ standard library. We hardcoded the mito::real to double for now to avoid this compilation error. We should figure out why this is happening and fix it later.