discretization: major redesign of the components discrete system / linear system / solver / matrix solver
In the new design, the linear system of equations is owned by the {DiscreteSystem} class and the matrix solver is owned by the {LinearSolver} class.
Classes {DiscreteSystem} and {LinearSolver} are unaware of implementation details such as the representation of the matrix system of equations and of the matrix solver (petsc vs. alternatives).

After chosing a (compatible) pair of matrix representation and matrix solver (e.g. {petsc::linear_system_t} and {petsc::ksp_t}), the user instantiates the {DiscreteSystem} first and then a {LinearSolver} on top of that {DiscreteSystem}. When the {LinearSolver} solves the linear system of equations (e.g. {PETScLinearSystem}) through a matrix solver (e.g. {PETScKrylovSolver}), it has the {DiscreteSystem} read the solution from the {PETScLinearSystem}.

After the {solve} call, the {LinearSolver} can be safely destroyed as the solution lies with the {DiscreteSystem}.