manifolds/geometry/quadrature: remove cell-level {parametrization} method from class {Manifold}
This method has been moved to {GeometricSimplex}. This way class {Manifold} doesn't need knowledge of lower level type definitions, such as the type of barycentric coordinates.
Also, this method in class {Manifold} was problematic in that it implied (without checking) that the cell passed in input to the method did belong to the manifold.